Fix same context to be used in different types of Notifications
The issue was found in Scenario module integration, when different
scenarios (new subscription, ending subscription) would use the same
context - because they all had subscription_id in the params.

This prevented later sent emails to be sent, because Notification
handlers (such as Mailer) check if the context was previously sent
and prevent same context to be sent again.

remp/crm#1515