Nette API: ParamsProcessor now extends NetteApi's ParamProcessor
`isError()` now returns bool only. In case you still need `isError()`
to return first encountered error, use temp method `hasError()` (which
is copy of previous method). But you'll need to fix this in future and
use `isError()` together with `getErrors()`.

This:

```php
$error = $paramsProcessor->hasError();
if ($error) {
    $response = new JsonResponse([
        'status' => 'error',
        'message' => 'Wrong input - ' . $error,
    ]);
}
```

Should be fixed to:

```php
if ($paramsProcessor->isError()) {
    $response = new JsonResponse([
        'status' => 'error',
        'code' => 'invalid_request',
        'errors' => $paramsProcessor->getErrors(),
    ]);
}
```

Unfortunately this is breaking change so we are not pushing this forward
right now.

remp/crm#333