Remove high priority of RefreshUserDataTokenHandler
No need to run this with higher priority after new subscription was
created. It is remnant of legacy code and currently it just creates
confusion.

Refresh of user date should run after all other handlers in case
subscription / subscription meta were updated by one of the event
handlers. So if any priority, it should be negative. Keeping it now
in same level as others.

remp/crm#3017